# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Ticket Booking API",
      "description": "One space portal for organizers and clients to provide and book tickets of a particular show/movie",
      "version": "1.0"
    },
    "produces": ["application/json"],
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
...
      "/api/user/bookings": {
        "get": {
            "tags": ["User Routes"],
            "summary": "Route for viewing all the bookings of the user.",
            "description": "Authentication header is required",
            "parameters": [{
                "in": "header",
                "name": "Authorization",
                "description": "Authorization Header",
                "required": "true",
                "type": "string"
              }],
            "responses": {}
        }
      },
      "/api/user/shows/{id}": {
        "get": {
            "tags": ["User Routes"],
            "summary": "Route for viewing details of a particular show.",
            "description": "All the slots are listed here.",
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "ID of the show",
                "required": "true",
                "type": "string"
            }],
            "responses": {}
        }
      },
      "/api/user/shows/slots/{id}": {
        "get": {
            "tags": ["User Routes"],
            "summary": "Route for viewing slots for a particular show.",
            "description": "All the details of selected slot are listed here.",
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "ID of the slot",
                "required": "true",
                "type": "string"
            }],
            "responses": {}
        },
        "post": {
            "tags": ["User Routes"],
            "summary": "Route for booking seats.",
            "description": "Route for booking a particular seat",
            "parameters": [{
                "name": "id",
                "in": "path",
                "description": "ID of the slot",
                "required": "true",
                "type": "string"
            },{
                "in": "header",
                "name": "Authorization",
                "description": "Authorization Header",
                "required": "true",
                "type": "string"
            },{
                "in": "body",
                "name": "body",
                "description": "Create a provider.",
                "schema": {
                    "$ref": "#/components/buyTickets"
                }
            }],
            "responses": {}
        }
      }
    },
...
 }